keep the diamond daa file

+++ graftm/diamond.py
@@ -10,7 +10,7 @@ class Diamond:
         self._threads = threads
         self._evalue = evalue
         
-    def run(self, input_sequence_file, input_sequence_type):
+    def run(self, input_sequence_file, input_sequence_type, daa_file_basename=None):
         '''Run input sequences in either blastp or blastx mode against the
         database specified in __init__.
             
@@ -34,31 +34,37 @@ class Diamond:
         else:
             raise Exception("Programming error")
         
-        with tempfile.NamedTemporaryFile(prefix='graftm_diamond') as t:
-            for c in ['-k 1',
-                      "-d",
-                        self._database,
-                        "-q",
-                        "%s" % input_sequence_file,
-                        "-a",
-                        t.name]:
-                cmd_list.append(c)
-            if self._threads:
-                cmd_list.append("--threads")
-                cmd_list.append(str(self._threads))
-            if self._evalue:
-                cmd_list.append("--evalue")
-                cmd_list.append(str(self._evalue))
+        basename = daa_file_basename
+        if basename is None:
+            with tempfile.NamedTemporaryFile(prefix='graftm_diamond') as t:
+                # we are just stealing the name, don't need the file itself
+                basename = t.name
             
-            cmd = ' '.join(cmd_list)
-            extern.run(cmd)
-            
-            daa_name = "%s.daa" % t.name
-            res = DiamondSearchResult.import_from_daa_file(daa_name)
-            
-            # Diamond makes an extra file, need to remove this specifically,
-            # not just the tempfile
+        for c in ['-k 1',
+                  "-d",
+                    self._database,
+                    "-q",
+                    "%s" % input_sequence_file,
+                    "-a",
+                    basename]:
+            cmd_list.append(c)
+        if self._threads:
+            cmd_list.append("--threads")
+            cmd_list.append(str(self._threads))
+        if self._evalue:
+            cmd_list.append("--evalue")
+            cmd_list.append(str(self._evalue))
+        
+        cmd = ' '.join(cmd_list)
+        extern.run(cmd)
+        
+        daa_name = "%s.daa" % basename
+        res = DiamondSearchResult.import_from_daa_file(daa_name)
+        
+        if daa_file_basename is None:
+            # Diamond makes an extra file, need to remove this
             os.remove(daa_name)
+            
         return res
